feature/#44 Test case for LIC 12 partial compete
a true test in CMVtest.java

+++ my-app/src/main/java/com/mycompany/app/LICutils.java
@@ -2,7 +2,7 @@ package com.mycompany.app;
 
 public class LICutils {
 
-    public static Double dist(int x1, int y1, int x2, int y2) {
+    public static Double dist(double x1, double y1, double x2, double y2) {
         return Math.sqrt((x1 - x2) * (x1 - x2) + (y1 - y2) * (y1 - y2));
     }
